+++ CHANGELOG.md
@@ -13,6 +13,12 @@ Types of changes are:
 6. Security - For security features
 
 ## Unreleased
+### Added
+- Created dom.ts with `breakpoints`, `isClient`
+- Added `scroll lock` to drawer and modal component
+
+### Fixed
+- Updated `Paganiation` resposiveness
 
 ## 1.0.26 - 2025-03-12
 ### Added